+++ sync-for-expenses/pkg/models/operations/createexpensetransaction.go
@@ -3,7 +3,7 @@
 package operations
 
 import (
-	"github.com/codatio/client-sdk-go/sync-for-expenses/v3/pkg/models/shared"
+	"github.com/codatio/client-sdk-go/sync-for-expenses/v4/pkg/models/shared"
 	"net/http"
 )
 
@@ -32,8 +32,6 @@ type CreateExpenseTransactionResponse struct {
 	ContentType string
 	// OK
 	CreateExpenseResponse *shared.CreateExpenseResponse
-	// The request made is not valid.
-	ErrorMessage *shared.ErrorMessage
 	// HTTP response status code for this operation
 	StatusCode int
 	// Raw HTTP response; suitable for custom response parsing
@@ -52,13 +50,6 @@ func (o *CreateExpenseTransactionResponse) GetCreateExpenseResponse() *shared.Cr
 		return nil
 	}
 	return o.CreateExpenseResponse
-}
-
-func (o *CreateExpenseTransactionResponse) GetErrorMessage() *shared.ErrorMessage {
-	if o == nil {
-		return nil
-	}
-	return o.ErrorMessage
 }
 
 func (o *CreateExpenseTransactionResponse) GetStatusCode() int {