+++ sync-for-expenses/pkg/models/operations/uploadexpenseattachment.go
@@ -3,13 +3,13 @@
 package operations
 
 import (
-	"github.com/codatio/client-sdk-go/sync-for-expenses/v3/pkg/models/shared"
+	"github.com/codatio/client-sdk-go/sync-for-expenses/v4/pkg/models/shared"
 	"net/http"
 )
 
 type UploadExpenseAttachmentRequestBody struct {
-	Content     []byte `multipartForm:"content"`
-	RequestBody string `multipartForm:"name=requestBody"`
+	Content  []byte `multipartForm:"content"`
+	FileName string `multipartForm:"name=fileName"`
 }
 
 func (o *UploadExpenseAttachmentRequestBody) GetContent() []byte {
@@ -19,11 +19,11 @@ func (o *UploadExpenseAttachmentRequestBody) GetContent() []byte {
 	return o.Content
 }
 
-func (o *UploadExpenseAttachmentRequestBody) GetRequestBody() string {
+func (o *UploadExpenseAttachmentRequestBody) GetFileName() string {
 	if o == nil {
 		return ""
 	}
-	return o.RequestBody
+	return o.FileName
 }
 
 type UploadExpenseAttachmentRequest struct {
@@ -69,8 +69,6 @@ type UploadExpenseAttachmentResponse struct {
 	Attachment *shared.Attachment
 	// HTTP response content type for this operation
 	ContentType string
-	// The request made is not valid.
-	ErrorMessage *shared.ErrorMessage
 	// HTTP response status code for this operation
 	StatusCode int
 	// Raw HTTP response; suitable for custom response parsing
@@ -89,13 +87,6 @@ func (o *UploadExpenseAttachmentResponse) GetContentType() string {
 		return ""
 	}
 	return o.ContentType
-}
-
-func (o *UploadExpenseAttachmentResponse) GetErrorMessage() *shared.ErrorMessage {
-	if o == nil {
-		return nil
-	}
-	return o.ErrorMessage
 }
 
 func (o *UploadExpenseAttachmentResponse) GetStatusCode() int {