+++ sync-for-expenses/pkg/models/shared/accountmappinginfo.go
@@ -44,24 +44,24 @@ func (e *AccountMappingInfoAccountType) UnmarshalJSON(data []byte) error {
 	}
 }
 
-type AccountMappingInfoValidTransactionTypes string
+type ValidTransactionTypes string
 
 const (
-	AccountMappingInfoValidTransactionTypesPayment       AccountMappingInfoValidTransactionTypes = "Payment"
-	AccountMappingInfoValidTransactionTypesRefund        AccountMappingInfoValidTransactionTypes = "Refund"
-	AccountMappingInfoValidTransactionTypesReward        AccountMappingInfoValidTransactionTypes = "Reward"
-	AccountMappingInfoValidTransactionTypesChargeback    AccountMappingInfoValidTransactionTypes = "Chargeback"
-	AccountMappingInfoValidTransactionTypesTransferIn    AccountMappingInfoValidTransactionTypes = "TransferIn"
-	AccountMappingInfoValidTransactionTypesTransferOut   AccountMappingInfoValidTransactionTypes = "TransferOut"
-	AccountMappingInfoValidTransactionTypesAdjustmentIn  AccountMappingInfoValidTransactionTypes = "AdjustmentIn"
-	AccountMappingInfoValidTransactionTypesAdjustmentOut AccountMappingInfoValidTransactionTypes = "AdjustmentOut"
+	ValidTransactionTypesPayment       ValidTransactionTypes = "Payment"
+	ValidTransactionTypesRefund        ValidTransactionTypes = "Refund"
+	ValidTransactionTypesReward        ValidTransactionTypes = "Reward"
+	ValidTransactionTypesChargeback    ValidTransactionTypes = "Chargeback"
+	ValidTransactionTypesTransferIn    ValidTransactionTypes = "TransferIn"
+	ValidTransactionTypesTransferOut   ValidTransactionTypes = "TransferOut"
+	ValidTransactionTypesAdjustmentIn  ValidTransactionTypes = "AdjustmentIn"
+	ValidTransactionTypesAdjustmentOut ValidTransactionTypes = "AdjustmentOut"
 )
 
-func (e AccountMappingInfoValidTransactionTypes) ToPointer() *AccountMappingInfoValidTransactionTypes {
+func (e ValidTransactionTypes) ToPointer() *ValidTransactionTypes {
 	return &e
 }
 
-func (e *AccountMappingInfoValidTransactionTypes) UnmarshalJSON(data []byte) error {
+func (e *ValidTransactionTypes) UnmarshalJSON(data []byte) error {
 	var v string
 	if err := json.Unmarshal(data, &v); err != nil {
 		return err
@@ -82,10 +82,10 @@ func (e *AccountMappingInfoValidTransactionTypes) UnmarshalJSON(data []byte) err
 	case "AdjustmentIn":
 		fallthrough
 	case "AdjustmentOut":
-		*e = AccountMappingInfoValidTransactionTypes(v)
+		*e = ValidTransactionTypes(v)
 		return nil
 	default:
-		return fmt.Errorf("invalid value for AccountMappingInfoValidTransactionTypes: %v", v)
+		return fmt.Errorf("invalid value for ValidTransactionTypes: %v", v)
 	}
 }
 
@@ -99,7 +99,7 @@ type AccountMappingInfo struct {
 	// Name of the account as it appears in the companies accounting software.
 	Name *string `json:"name,omitempty"`
 	// Supported transaction types for the account.
-	ValidTransactionTypes []AccountMappingInfoValidTransactionTypes `json:"validTransactionTypes,omitempty"`
+	ValidTransactionTypes []ValidTransactionTypes `json:"validTransactionTypes,omitempty"`
 }
 
 func (o *AccountMappingInfo) GetAccountType() *AccountMappingInfoAccountType {
@@ -130,7 +130,7 @@ func (o *AccountMappingInfo) GetName() *string {
 	return o.Name
 }
 
-func (o *AccountMappingInfo) GetValidTransactionTypes() []AccountMappingInfoValidTransactionTypes {
+func (o *AccountMappingInfo) GetValidTransactionTypes() []ValidTransactionTypes {
 	if o == nil {
 		return nil
 	}