+++ sync-for-expenses/RELEASES.md
@@ -383,3 +383,13 @@ Based on:
 - [go v3.2.0] sync-for-expenses
 ### Releases
 - [Go v3.2.0] https://github.com/codatio/client-sdk-go/releases/tag/sync-for-expenses/v3.2.0 - sync-for-expenses
+
+## 2023-11-27 01:25:16
+### Changes
+Based on:
+- OpenAPI Doc prealpha https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Sync-Expenses.yaml
+- Speakeasy CLI 1.122.0 (2.195.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [go v4.0.0] sync-for-expenses
+### Releases
+- [Go v4.0.0] https://github.com/codatio/client-sdk-go/releases/tag/sync-for-expenses/v4.0.0 - sync-for-expenses